Bulk merge several contributions.
- Add a few more HTML5 and Intercooler attributes.

- Add several Bootstrap attributes.

- Add the HTML5 template element.

- Indent option elements within select elements.

- Store element definitions inside element objects.

- Store attribute type functions inside attribute objects.

The last two changes make the handling of "dummy" elements and
"extended" attributes much cleaner and eliminate the need to
check for elements being valid in the DTD (which was really a
holdover from the time when we supported more than just the
Strict DTD).

One observable benefit of the new design is that the HTML5
autofocus and required attributes are now rendered properly as
booleans. (The same cannot be done for Bootstrap and Intercooler
boolean attributes because they use the string "true" instead of
the standard HTML style.)